Rename Temp to Main
