+++ src/App.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { RecoilRoot } from "recoil";
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
@@ -7,7 +6,6 @@ import Intro from '@/pages/Intro';
 import Login from '@/pages/Login';
 import SignUp from '@/pages/SignUp';
 import Main from '@/pages/Main';
-import Temp from '@/pages/Temp';
 
 const queryClient = new QueryClient();
 
@@ -21,7 +19,6 @@ function App() {
           <Route path="/login/:provider" element={ <Login /> }/>
           <Route path="/sign-up" element={<SignUp />} />
           <Route path="/main" element={<Main />} />
-          <Route path="/temp" element={<Temp />} />
         </Routes>
       </BrowserRouter>
       </QueryClientProvider>
